OutOfBounds when checking MetaDataRenderer axes
The MetaData-(and Grid-)Renderer always returns an empty axesList.
Because of that, even if the code before adds 2 axes if they are
not available, subList(0,2) throws IndexOutOfBounds.
Fixed by limiting the stream to the first 2 entries with limit(2).

Open points:
- because of the behaviour of the MetaDataRenderer, the axes are added
every render cycle to a newly created list which is immediately dropped
afterwards. Add special case and throw OperationNotSupported on
MetaDataSetRenderer.getAxes() ? Move getAxes from RendererInterface to
DataSetManagement (which is the only implementation at the moment anyway,
but would force custom renderers to extend it :/)?
- while often the x and y axis are the first 2 axes in the list, this
is not enforced anywhere, so the z axis could also be the first axis or
there could be 2 x axes in the renderer.